+++ crates/reinhardt-utils/CHANGELOG.md
@@ -7,6 +7,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.5](https://github.com/kent8192/reinhardt-web/compare/reinhardt-utils@v0.3.4...reinhardt-utils@v0.3.5) - 2026-08-02
+
+### Maintenance
+
+- update Cargo.toml dependencies
+
+## [0.3.4](https://github.com/kent8192/reinhardt-web/compare/reinhardt-utils@v0.3.3...reinhardt-utils@v0.3.4) - 2026-07-30
+
+### Maintenance
+
+- update Cargo.toml dependencies
+
+## [0.3.3](https://github.com/kent8192/reinhardt-web/compare/reinhardt-utils@v0.3.2...reinhardt-utils@v0.3.3) - 2026-07-28
+
+### Maintenance
+
+- update Cargo.toml dependencies
+
 ## [0.3.0](https://github.com/kent8192/reinhardt-web/compare/reinhardt-utils@v0.2.0...reinhardt-utils@v0.3.0) - 2026-06-28
 
 Stable release of `reinhardt-utils` for the Reinhardt 0.3.0 line. This